requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==46.0.7
exceptiongroup==1.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
joblib==1.5.3
lxml==6.1.0
numpy==2.2.6
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.7.1
python-docx==1.2.0
python-multipart==0.0.22
scikit-learn==1.7.2
scipy==1.15.3
starlette==0.52.1
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
